+++ src/components/dialog.tsx
@@ -1,4 +1,6 @@
-import React, { useState } from 'react';
+import React, { useReducer } from 'react';
+
+import { ImgButton } from '../compounds/imgbutton';
 
 export function Dialog({
    id,
@@ -14,9 +16,7 @@ export function Dialog({
    title: string;
    trigger: JSX.Element;
 }) {
-   let [o, setOpen] = useState(open);
-
-   let openClose = () => setOpen((old: boolean) => !old);
+   let [o, openClose] = useReducer((old: boolean) => !old, open);
 
    return (
       <div id={id} className='dialog' {...attrs}>
@@ -26,7 +26,7 @@ export function Dialog({
          <dialog open={o}>
             <div className='header'>
                <h1>{title}</h1>
-               <img src='https://img.icons8.com/emoji/96/FAB005/cross-mark-emoji.png' alt='Close' onClick={() => openClose()} />
+               <ImgButton alt='Close' click={() => openClose()} src='https://img.icons8.com/fluency/48/null/delete-sign.png' />
             </div>
             <div className='children'>{children}</div>
          </dialog>